Moved (and tweaked) Tool.list() test into test-tool-model.js
